refactor(frontend): split vllm-frontend lib.rs into bridge/lora/wire modules
